+++ mayan/apps/appearance/apps.py
@@ -6,6 +6,7 @@ from mayan.apps.common.apps import MayanAppConfig
 
 
 class AppearanceApp(MayanAppConfig):
+    has_tests = True
     name = 'mayan.apps.appearance'
     verbose_name = _('Appearance')